:sparkles: feat: integration with the editing route

+++ server/src/api/v1/dto/transaction_dto.go
@@ -5,7 +5,7 @@ import "time"
 type TransactionRequestDto struct {
 	Title       string  `json:"title"`
 	Description *string `json:"description,omitempty"`
-	Amount      int64   `json:"amount"`
+	Amount      float64 `json:"amount"`
 	Category    string  `json:"category"`
 	Type        string  `json:"type"`
 }
@@ -14,7 +14,7 @@ type TransactionResponseDto struct {
 	ID          uint      `json:"id"`
 	Title       string    `json:"title"`
 	Description *string   `json:"description,omitempty"`
-	Amount      int64     `json:"amount"`
+	Amount      float64   `json:"amount"`
 	Category    string    `json:"category"`
 	Type        string    `json:"type"`
 	CreatedAt   time.Time `json:"created_at"`
